Added gradients to the plot and directional derivative

+++ tcav/collect_from_model/collect_from_model.py
@@ -6,6 +6,7 @@ import numpy as np
 import PIL.Image
 from multiprocessing import dummy as multiprocessing
 import math
+from datetime import datetime
 
 def create_session():
     graph = tf.Graph()
@@ -81,7 +82,7 @@ def main():
     bottleneck =  'mixed4d'   
     max_examples = 100
     sess = create_session()
-
+    wd = '/media/adri/Spare/acts_grads/'
     classes = None
     with open('../../concept-vis/data/classes.json') as f:
         classes = json.loads(f.read())
@@ -96,30 +97,37 @@ def main():
                                             GRAPH_PATH,
                                             LABEL_PATH)
 
-    for item in classes:
-        imgs = os.listdir(images_path + item['id'] + '/images')
-        filenames = [images_path + item['id'] + '/images/' + x for x in imgs] 
-        examples = load_images_from_files(filenames)
+    for i, item in enumerate(classes):
+        if not os.path.isfile(wd + item['id'] + "_sorted_grads"):
+            start_time = datetime.now()
 
-        acts_all = []
-        grads_all = []
-        for batch in range(math.ceil(len(imgs)/max_examples)):
-             
-            acts = mymodel.run_examples(examples[batch*max_examples:(batch+1)*max_examples], bottleneck)
-            print(np.shape(acts))
-            acts = mymodel.reshape_activations(acts).squeeze()
-            class_id = mymodel.label_to_id(item['class'].split(',')[0])
-            grad = mymodel.get_gradient(acts, [class_id], bottleneck)
-            print(np.shape(grad))
-            reshaped = np.reshape(grad, -1)
-            acts_all.append(acts)
-            grads_all.append(grad)
-       
-            with tf.gfile.Open('./'+item['id']+'_acts', 'w') as f:
-              np.save(f, acts_all, allow_pickle=False) 
-            with tf.gfile.Open('./'+item['id']+'_grads', 'w') as f:
-              np.save(f, grads_all, allow_pickle=False) 
-        break
+            imgs = sorted(os.listdir(images_path + item['id'] + '/images'))
+            filenames = [images_path + item['id'] + '/images/' + x for x in imgs] 
+            examples = load_images_from_files(filenames)
+            with open(wd + item['id']+ '_names','w') as f:
+                f.write('\n'.join(imgs))
+            acts_all = []
+            grads_all = []
+            for batch in range(math.ceil(len(imgs)/max_examples)):
+                try:     
+                    acts = mymodel.run_examples(examples[batch*max_examples:(batch+1)*max_examples], bottleneck)
+                    acts = mymodel.reshape_activations(acts).squeeze()
+                    class_id = mymodel.label_to_id(item['class'].split(',')[0])
+                    grad = mymodel.get_gradient(acts, [class_id], bottleneck)
+                    reshaped = np.reshape(grad, -1)
+                    acts_all.append(acts)
+                    grads_all.append(grad)
+               
+                    with tf.gfile.Open(wd+item['id']+'_sorted_acts', 'w') as f:
+                      np.save(f, acts_all, allow_pickle=False) 
+                    with tf.gfile.Open(wd+item['id']+'_sorted_grads', 'w') as f:
+                      np.save(f, grads_all, allow_pickle=False) 
+                except Exception as e:
+                    print("Problem with class: {} with id: {}".format(item["class"],item['id']))
+                    print(e)
+            delta = datetime.now() - start_time 
+            print("Class: {1} activations and gradients collected (id: {0}) took: {2} | remaining time estimate: {3}".format(item['id'],item['class'],delta, delta*(len(classes)-i+1)))
+
 
     print("DONE!")
